BF: fix path issue
The test assumed that the cli files are available on the path. Since we want to
execute the latest version from our git repository we adapt the path.
We don't do this for all tests, since simply fixing the path does not make them
work. There are other issues that need to be fixed first.