Added pickle support for PsiInference objects
There are two attributes (_data and _pmf) that are swig-objects and, therefore,
not pickle-able.  However, these can be reconstructed after pickling using the
swignifit.util functions.  The other issue was the MCMCPars subclass of Dict
that was defined inside a function (pygibssit.py).  By moving this outside of
the function body, pickle was able to reconstruct the original class.

Signed-off-by: Ingo Fründ <ingo.fruend@googlemail.com>